Compute engineName based on extension and all available engines
This should be compatible with previous version but it won't set engine
to python based on path of test.py if python engine not on classpath.

This fixes a few logging bugs and bug in one of the constructors that is
from my previous commit (calling getEngineName() to set enginename)

This also avoids checking for file and classpath contents every time the
script is run, not type of script is determined when scriptFile is set
in constructor or via setter.

Changes to scripts will be still be picked up b/c script files or
classpath
resources are still re-read each time script runs, as before.

Log warning that engine can't be found when engineName is set by user
(if it isn't the classpath)