Allow ScriptEnginePersonAttributeDao to pass attributes through to script.
Change should not break existing scripts.

This changes MultivaluedPersonAttributeUtils.toMultivaluedMap
to not wrap a List in a List and removes similar method that
did same thing.
I would consider this fixing a bug and while it could conceivably
be a breaking change, this is 2.0 version.

GroovyPersonAttributeDao already has ability to get attributes
in script invocation depending on what method the script implements.
GroovyPersonAttributeDao should probably be deprecated
in favor of ScriptEnginePersonAttributeDao.