remove all deprecated methods for 2.0 release (#120)
* remove all deprecated methods

update tests that used old methods to use nearest replacement

included the method name in caching key so return type
of different methods won't be result in runtime error by pulling
object out of cache and returning it for method that
should return a different type (e.g. Set<IPersonAttributes> vs IPersonAttributes

* Allow ScriptEnginePersonAttributeDao to pass attributes through to script.

Change should not break existing scripts.

This changes MultivaluedPersonAttributeUtils.toMultivaluedMap
to not wrap a List in a List and removes similar method that
did same thing.
I would consider this fixing a bug and while it could conceivably
be a breaking change, this is 2.0 version.

GroovyPersonAttributeDao already has ability to get attributes
in script invocation depending on what method the script implements.
GroovyPersonAttributeDao should probably be deprecated
in favor of ScriptEnginePersonAttributeDao.